2nd-place/requirements.txt
rasterio           == 1.3.4
matplotlib         == 3.6.2
pandas             == 1.5.1
tqdm               == 4.64.1
omegaconf          == 2.2.3
einops             == 0.6.0
volumentations-3D  == 1.0.4
monai              == 1.0.1
pillow             == 9.3.0
piqa               == 1.2.2